+++ Tests/Unit/EmogrifierTest.php
@@ -642,6 +642,15 @@ class EmogrifierTest extends \PHPUnit_Framework_TestCase
             'attribute value selector SPAN[title$] matches element with an attribute value that end that word and contains other words' => [
                 'span[title$="dias"] {' . $styleRule . '} ', '#<span title="buenas dias" ' . $styleAttribute . '>#'
             ],
+            'attribute value selector SPAN[title|] matches element with an attribute value that is exactly that word' => [
+                'span[title|="bonjour"] {' . $styleRule . '} ', '#<span title="bonjour" ' . $styleAttribute . '>#'
+            ],
+            'attribute value selector SPAN[title|] matches element with an attribute value with two words' => [
+                'span[title|="buenas dias"] {' . $styleRule . '} ', '#<span title="buenas dias" ' . $styleAttribute . '>#'
+            ],
+            'attribute value selector SPAN[title|] matches element with an attribute value with two words with hypen' => [
+                'span[title|="avez"] {' . $styleRule . '} ', '#<span title="avez-vous" ' . $styleAttribute . '>#'
+            ],
             'attribute value selector SPAN[title] does not match element with other attribute value'
                 => ['span[title="bonjour"] {' . $styleRule . '} ', '#<span title="buenas dias">#'],
             'attribute value selector SPAN[title] does not match element without any attributes'
@@ -653,14 +662,14 @@ class EmogrifierTest extends \PHPUnit_Framework_TestCase
             'P:first-child does not match middle child'
                 => ['p:first-child {' . $styleRule . '} ', '#<p class="p-2">#'],
             'P:first-child does not match last child'
-                => ['p:first-child {' . $styleRule . '} ', '#<p class="p-4">#'],
+                => ['p:first-child {' . $styleRule . '} ', '#<p class="p-5">#'],
             'P:last-child does not match first child' => ['p:last-child {' . $styleRule . '} ', '#<p class="p-1">#'],
             'P:last-child does not match middle child'
-                => ['p:last-child {' . $styleRule . '} ', '#<p class="p-2">#'],
+                => ['p:last-child {' . $styleRule . '} ', '#<p class="p-3">#'],
             'P:last-child matches last child'
-                => ['p:last-child {' . $styleRule . '} ', '#<p class="p-4" style="' . $styleRule . '">#'],
+                => ['p:last-child {' . $styleRule . '} ', '#<p class="p-5" style="' . $styleRule . '">#'],
             'DIV:last-child does not match last child with mismatching tag'
-                => ['div:last-child {' . $styleRule . '} ', '#<p class="p-4">#'],
+                => ['div:last-child {' . $styleRule . '} ', '#<p class="p-5">#'],
         ];
     }
 
@@ -680,7 +689,8 @@ class EmogrifierTest extends \PHPUnit_Framework_TestCase
             '    <p class="p-1"><span>some text</span></p>' .
             '    <p class="p-2"><span title="bonjour">some</span> text</p>' .
             '    <p class="p-3"><span title="buenas dias">some</span> more text</p>' .
-            '    <p class="p-4"><span title="buenas dias bom dia">some</span> more text</p>' .
+            '    <p class="p-4"><span title="avez-vous">some</span> more text</p>' .
+            '    <p class="p-5"><span title="buenas dias bom dia">some</span> more text</p>' .
             '  </body>' .
             '</html>';
         $this->subject->setHtml($html);